# manager/swagger.yml
openapi: 3.0.0
info:
  title: Hash Cracker API
  description: |
    Распределённая система для взлома MD5 хешей методом перебора.
    
    ## Архитектура
    - **Менеджер** (порт 8000) - координирует работу воркеров
    - **Воркеры** (порт 9000) - выполняют непосредственный перебор
    
    ## Особенности
    - Автоматическое распределение задач между воркерами
    - Отслеживание прогресса выполнения
    - Механизм восстановления при сбоях воркеров
    - Дедлайн 20 секунд на ответ от воркера
    - Максимум 3 попытки перезапуска упавшего воркера
  version: 1.0.0
  contact:
    name: Hash Cracker Team
    email: support@hashcracker.local
servers:
  - url: http://manager:8000
    description: Менеджер (внутренняя сеть Docker)
  - url: http://localhost:8000
    description: Менеджер (локальный доступ)
...
paths:
...
  /internal/api/worker/hash/crack/task:
    post:
      tags:
        - Internal API
      summary: Отправить задачу воркеру
      description: |
        Менеджер отправляет часть диапазона для перебора конкретному воркеру.
        **Внутренний эндпоинт воркера, доступен только менеджеру.**
      operationId: assignWorkerTask
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerTaskAssignment'
      responses:
        '200':
          description: Задача успешно обработана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerTaskResult'
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка при выполнении
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerTaskResult'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Описание ошибки
          example: "word is required"
    
    # Статусы задачи
...
    WorkerTaskAssignment:
      type: object
      required:
        - word
        - idx_start
        - idx_end
        - task_id
        - worker_id
      properties:
        word:
          type: string
          description: MD5 хеш для взлома
        idx_start:
          type: integer
          description: Начальный индекс для перебора (включительно)
          example: 0
        idx_end:
          type: integer
          description: Конечный индекс для перебора (исключительно)
          example: 1000
        task_id:
          type: string
          format: uuid
          description: ID задачи
        worker_id:
          type: string
          description: ID воркера
        last_processed_index:
          type: integer
          description: Индекс, с которого нужно продолжить (для восстановления)
        found_words:
          type: array
          items:
            type: string
          description: Уже найденные слова (для восстановления)
    WorkerTaskResult:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
          example: "ok"
        worker:
          type: string
          description: Имя хоста воркера
          example: "worker_1"
        worker_id:
          type: string
          description: ID воркера
        task_id:
          type: string
        received_word:
          type: array
          items:
            type: string
          description: Найденные слова
        range_processed:
          type: string
          description: Обработанный диапазон
          example: "[0, 1000)"
        last_processed_index:
          type: integer
        total_checked:
          type: integer
        error:
          type: string
          description: Описание ошибки (при status=error)
...